% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{ANOVA Tables using the F distribution for glmmTMB models}
\usage{
anova(model, type = 3, ddf = "inner-outer", contr_sum = TRUE)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}

\item{type}{type of test, \code{II, III, 2, 3}. Roman numerals are equivalent
to the corresponding Arabic numerals.}

\item{ddf}{the method for computing the denominator degrees of freedom and F-statistics.
ddf = \code{inner-outer} primarily corresponds to classic \code{\link{nlme}}
denominator degrees of freedom calculations. \code{containment} corresponds to the SAS
default degrees of freedom calculations for generalized linear mixed models.}

\item{contr_sum}{a \code{boolean} indicating whether to apply 'sum to zero'
contrasts when performing the ANOVA calculation. Default is \code{TRUE}.}
}
\value{
a \code{data.frame}
}
\description{
The \code{anova} function will perform an analysis of variance on glmmTMB
    objects using the F distribution. Denominator degrees of freedom calculations are
    performed using either the \code{\link{nlme}} \code{inner-outer} or \code{containment}
    methods.
}
