% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aov_formula_writer}
\alias{aov_formula_writer}
\title{AOV slope and slope/intercept models equation developer}
\usage{
aov_formula_writer(model)
}
\arguments{
\item{model}{}
}
\value{
a \code{\link{formula}}
}
\description{
AOV slope and slope/intercept models equation developer
}
\keyword{internal}
