% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{base_aov_dfs}
\alias{base_aov_dfs}
\title{Performs basic ANOVA degrees of freedom calculations for input into downstream
functions.}
\usage{
base_aov_dfs(model)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}
}
\value{
a \code{\link{data.frame}}
}
\description{
Performs basic ANOVA degrees of freedom calculations for input into downstream
functions.
}
\keyword{internal}
