% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{best_rules}
\alias{best_rules}
\title{Identify the Most Appropriate Rules to Apply to each Fixed Effect Term}
\usage{
best_rules(model)
}
\arguments{
\item{model}{}
}
\value{
a \code{\link{data.frame}}
}
\description{
Identify the Most Appropriate Rules to Apply to each Fixed Effect Term
}
\keyword{internal}
