% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment.R
\name{containment_aov}
\alias{containment_aov}
\title{Containment degrees of freedom calculation}
\usage{
containment_aov(model = model, type = type)
}
\arguments{
\item{type}{type type of test, \code{II, III, 2, 3}. Roman numerals are equivalent
to the corresponding Arabic numerals.}

\item{a}{\code{\link{glmmTMB}} model object}
}
\value{
a \code{data.frame}
}
\description{
Containment degrees of freedom calculation
}
\keyword{internal}
