% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment.R
\name{containment_summary}
\alias{containment_summary}
\title{Containment Parameter Estimates Table Calculations}
\usage{
containment_summary(model = model)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}
}
\value{
a \code{data.frame}
}
\description{
Containment Parameter Estimates Table Calculations
}
