% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_aov_dfs.R
\name{data_aov_dfs}
\alias{data_aov_dfs}
\title{Performs basic ANOVA degrees of freedom calculations for input into downstream
functions with a formula and data supplied separately. This useful for random
slope / slope & int models}
\usage{
data_aov_dfs(formula, data)
}
\arguments{
\item{formula}{\code{\link{formula}}}

\item{data}{a \code{\link{data.frame}}}
}
\value{
a \code{\link{data.frame}}
}
\description{
Performs basic ANOVA degrees of freedom calculations for input into downstream
functions with a formula and data supplied separately. This useful for random
slope / slope & int models
}
\keyword{internal}
