% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_outer.R
\name{inner_outer_aov}
\alias{inner_outer_aov}
\title{Inner-Outer ANOVA and Degrees of Freedom Calculations}
\usage{
inner_outer_aov(model = model, type = type)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}

\item{type}{type of test, \code{II, III, 2, 3}. Roman numerals are equivalent
to the corresponding Arabic numerals.}
}
\value{
a \code{data.frame}
}
\description{
Inner-Outer ANOVA and Degrees of Freedom Calculations
}
\keyword{internal}
