% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_outer.R
\name{inner_outer_summary}
\alias{inner_outer_summary}
\title{Inner-Outer Parameter Estimates Table Calculations}
\usage{
inner_outer_summary(model = model)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}
}
\value{
a \code{data.frame}
}
\description{
Inner-Outer Parameter Estimates Table Calculations
}
