% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{model_properties}
\alias{model_properties}
\title{Define Model Properties for Various Terms}
\usage{
model_properties(model)
}
\arguments{
\item{model}{}
}
\value{
a \code{\link{data.frame}}
}
\description{
Define Model Properties for Various Terms
}
\keyword{internal}
