% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.R
\name{nlme_aov}
\alias{nlme_aov}
\title{nlme ANOVA and Degrees of Freedom Calculations}
\usage{
nlme_aov(model = model, type = type)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}

\item{type}{type of test, \code{II, III, 2, 3}. Roman numerals are equivalent
to the corresponding Arabic numerals.}
}
\value{
a \code{data.frame}
}
\description{
nlme ANOVA and Degrees of Freedom Calculations
}
