% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.R
\name{nlme_summary}
\alias{nlme_summary}
\title{nlme Parameter Estimates Table Calculations}
\usage{
nlme_summary(model = model)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}
}
\value{
a \code{data.frame}
}
\description{
nlme Parameter Estimates Table Calculations
}
